requirements.txt
torch
torchvision
Pillow
numpy
pytest
